requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
coloredlogs==15.0.1
crypto==1.4.1
cryptography==41.0.4
ddddocr==1.4.7
exceptiongroup==1.1.3
flatbuffers==23.5.26
h11==0.14.0
humanfriendly==10.0
idna==3.4
iniconfig==2.0.0
mpmath==1.3.0
Naked==0.1.32
numpy==1.25.2
onnxruntime==1.15.1
opencv-python-headless==4.8.0.76
outcome==1.2.0
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
protobuf==4.24.1
pyasn1==0.5.0
pycparser==2.21
pycryptodome==3.19.0
PyMySQL==1.1.0
pyreadline3==3.4.1
PySocks==1.7.1
pytest==7.4.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
selenium==4.11.2
shellescape==3.8.1
sniffio==1.3.0
sortedcontainers==2.4.0
sympy==1.12
tomli==2.0.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0

pip~=23.3.2
wheel~=0.38.4
Jinja2~=3.1.2
setuptools~=65.5.1
paramiko~=3.4.0